Comment on PR if CLA is not signed
The action will now comment on PR directly if any authors has not signed
the CLA.
Any rerun (commit or manual) will update the existing comment and
add/remove to the list of non signers.
If everyone has signed the CLA, the existing comment will be updated
again for one last time.
Finally, no comment is posted in the PR if everyone has signed the CLA.